backend/requirements.txt
django
django-restframework
django-cors-headers
djangorestframework
psycopg2
Pillow